requirements.txt
Django>=4.1.6
django-filter==22.1
djangorestframework==3.14.0
isort==5.12.0
flake8==3.7.8
Pillow==6.1.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pytz==2019.1
sqlparse==0.3.0
